fix(workspace): unblock build — TryFrom imports, ambiguous numerics, borrow-checker
Follow-up to 9fba0d14b ("fix(gui): refuse block_on on the main-thread
spawn queue and guard run-loop re-entry"). The deadlock-defense commit
left the wider workspace unable to build because of pre-existing
breakage in several crates. This commit clears all of it so the GUI
can actually be rebuilt and the assertion can ride.

* `frankenterm-term`, `mux`, `codec` (8 files) — restore explicit
  `use std::convert::TryFrom;` imports that were dropped at some point
  but are still required to compile against this Rust edition.
* `term/terminalstate/mod.rs` — type the DECRQCRA checksum accumulator
  as `u16` (the function already returns u16); also extract the
  EraseToEndOfLine range start into a binding so the parser stops
  reading the if-block + `..cols` as a statement-then-RangeTo and the
  match arms unify on `Range<usize>`.
* `mux/pane.rs` — type `offset: usize` in `xy_to_logical_x`; in
  `for_each_logical_line_mut` capture `physical_lines.len()` before
  the `.iter_mut()` loop so the immutable read doesn't conflict with
  the live `&mut`s sitting in `line_refs`.
* `frankenterm-core/fleet_memory_controller.rs` — `u128` doesn't
  `impl From<usize>` (usize is platform-dependent); route through
  `as u64`, which is loss-free on every supported target.
* `frankenterm-core/fleet_scrollback_coordinator.rs` — explicit
  `usize` types on `let mut updated/applied/targets = 0;` so
  `saturating_add` resolves.
* `frankenterm-core/wezterm.rs` — `attempt: usize` vs
  `self.retry_attempts: u32` mismatch in the retry comparison,
  resolved with `as usize`.

No behavior changes intended; the diffs are minimum-invasive
restorations of build correctness.

+++ crates/frankenterm-core/src/fleet_memory_controller.rs
@@ -1155,8 +1155,11 @@ fn ceil_mul_div_usize(multiplicand: usize, numerator: usize, denominator: usize)
         return 0;
     }
 
-    let product = u128::from(multiplicand).saturating_mul(u128::from(numerator));
-    let denominator = u128::from(denominator);
+    // u128 does not implement From<usize> (usize is platform-dependent),
+    // but it implements From<u64> and usize always fits in u64 on all
+    // supported targets. Cast through u64.
+    let product = u128::from(multiplicand as u64).saturating_mul(u128::from(numerator as u64));
+    let denominator = u128::from(denominator as u64);
     debug_assert!(denominator > 0);
 
     let rounded = product.div_ceil(denominator);
@@ -1164,8 +1167,11 @@ fn ceil_mul_div_usize(multiplicand: usize, numerator: usize, denominator: usize)
 }
 
 fn floor_mul_div_usize(multiplicand: usize, numerator: usize, denominator: usize) -> usize {
-    let product = u128::from(multiplicand).saturating_mul(u128::from(numerator));
-    let denominator = u128::from(denominator);
+    // u128 does not implement From<usize> (usize is platform-dependent),
+    // but it implements From<u64> and usize always fits in u64 on all
+    // supported targets. Cast through u64.
+    let product = u128::from(multiplicand as u64).saturating_mul(u128::from(numerator as u64));
+    let denominator = u128::from(denominator as u64);
     debug_assert!(denominator > 0);
 
     usize::try_from(product / denominator).unwrap_or(usize::MAX)
